+++ public/js/main.js
@@ -33,6 +33,7 @@ import { initWaf, openWaf } from './waf.js';
 import { initStatus, openStatus } from './status.js';
 import { initWordlist, openWordlist } from './wordlist.js';
 import { initIpObf, openIpObf } from './ipobf.js';
+import { initAttackHdr, openAttackHdr } from './attackhdr.js';
 import { initHistory, renderHistory } from './history.js';
 import { initIdentities, populateSelect } from './identities.js';
 import { initAttacks } from './attacks.js';
@@ -77,6 +78,7 @@ async function boot() {
   initStatus();
   initWordlist();
   initIpObf();
+  initAttackHdr();
   initHistory();
   initExplorer(loadEndpoint);
   initAttacks();
@@ -116,6 +118,7 @@ async function boot() {
     status: openStatus,
     wordlist: openWordlist,
     ipobf: openIpObf,
+    attackhdr: openAttackHdr,
     'clear-filters': clearFilters,
     shortcuts: () => ($('#shortcutsModal').hidden = false),
     about: () => {